Added capability to split input and restart directories. Previously model input was all read from restart_dir and then restarts were written to the same directory.
This change adds input_dir to the setup_nml namelist. By default, if
input_dir is not specified in the namelist it will be set to the same
as restart_dir, reproducing previous behaviour, and so making this
backwards compatible.

If input_dir is specified then all inputs will be read from the specified
directory, and restarts will be written to restart_dir.