libaccessom2 intercommunicator profiling support
This patch modifies CICE to support changes to libaccessom2 which allow
the user to disable intercommunicators, instead using MPI_COMM_WORLD to
exchange messages between models.  This work is primarily to support
limitations in the Score-P profiler.

Specific changed detailed below.

- accessom2%config_sync uses the new API, and requires an libaccessom2
  update.

- il_commatm has been removed from cpl_interface, since coupler%atm_root
  provides this information.

- Most of the work in prism_init has been removed are replaced with
  equivalent operations in coupler%init_begin.

- Explicit MPI send/recv calls now use ranks provided by coupler, rather
  than explicit sends to 0 over OASIS-provided intercommunicators.

- Similarly, data sends now happen on the zero rank of CICE, rather than
  checking if the atmosphere rank of the intercommunicator is zero.
  Functionally, there should be no difference, but beware the
  unexpected.

- Buildscripts for score-p have been added.

Further changes to cpl_interface are possible, since most of the MPI
information is provided by the coupler object.